WIP - Pull apart classes for declarations and expressions
